services/worker/requirements.txt
azure-monitor-opentelemetry
pika
httpx
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp-proto-http


